refactor: profile 클라이언트 컴포넌트 분리
